lavf: merge utils-mythtv.c back into utils.c
