+++ mythtv/external/FFmpeg/libavformat/Makefile
@@ -21,7 +21,6 @@ OBJS = allformats.o         \
        sdp.o                \
        url.o                \
        utils.o              \
-       utils-mythtv.o       \
 
 OBJS-$(HAVE_LIBC_MSVCRT)                 += file_open.o
